Refactor and review comments
