+++ service/response/response.go
@@ -17,7 +17,7 @@
  *
  */
 
-package utils
+package response
 
 import (
 	"encoding/json"
